Moved Guess and Check code into GaCDataMonitor.cs
