+++ Adaptive Alarm/Adaptive Alarm/DataMonitors/GuessAndCheckData.cs
